feat: 대기열 등록 후, 토큰을 발급하는 api 작성. (polling용) 유저 토큰을 통해 대기열 정보 조회 api 작성.
chore: 패키지 구조 수정
- Mock Controller, presentation dto 삭제 예정.
