--- src/main/java/hhplus/concert/api/dto/ConcertScheduleInfo.java
@@ -1,10 +1,0 @@
-package hhplus.concert.api.dto;
-
-import java.time.LocalDateTime;
-
-public record ConcertScheduleInfo(
-        Long scheduleId,
-        LocalDateTime concertAt,
-        LocalDateTime reservationAt
-) {
-}
