--- src/main/java/hhplus/concert/api/dto/ConcertSeatInfo.java
@@ -1,11 +1,0 @@
-package hhplus.concert.api.dto;
-
-import hhplus.concert.common.type.SeatStatus;
-
-public record ConcertSeatInfo(
-        Long seatId,
-        int seatNumber,
-        SeatStatus seatStatus,
-        int SeatPrice
-) {
-}
