--- src/main/java/hhplus/concert/api/dto/request/BalanceChargeRequest.java
@@ -1,6 +1,0 @@
-package hhplus.concert.api.dto.request;
-
-public record BalanceChargeRequest(
-        Long amount
-) {
-}
