--- src/main/java/hhplus/concert/api/dto/request/PaymentRequest.java
@@ -1,6 +1,0 @@
-package hhplus.concert.api.dto.request;
-
-public record PaymentRequest(
-        Long reservationId
-) {
-}
