--- src/main/java/hhplus/concert/api/dto/request/ReservationRequest.java
@@ -1,11 +1,0 @@
-package hhplus.concert.api.dto.request;
-
-import hhplus.concert.common.type.ReservationStatus;
-
-public record ReservationRequest(
-        Long userId,
-        Long concertId,
-        Long scheduleId,
-        ReservationStatus status
-) {
-}
