--- src/main/java/hhplus/concert/api/dto/response/BalanceResponse.java
@@ -1,7 +1,0 @@
-package hhplus.concert.api.dto.response;
-
-public record BalanceResponse(
-        Long userId,
-        Long currentAmount
-) {
-}
