--- src/main/java/hhplus/concert/api/dto/response/ConcertScheduleResponse.java
@@ -1,11 +1,0 @@
-package hhplus.concert.api.dto.response;
-
-import hhplus.concert.api.dto.ConcertScheduleInfo;
-
-import java.util.List;
-
-public record ConcertScheduleResponse(
-        Long concertId,
-        List<ConcertScheduleInfo> events
-) {
-}
