--- src/main/java/hhplus/concert/api/dto/response/ConcertSeatResponse.java
@@ -1,13 +1,0 @@
-package hhplus.concert.api.dto.response;
-
-import hhplus.concert.api.dto.ConcertSeatInfo;
-
-import java.time.LocalDateTime;
-import java.util.List;
-
-public record ConcertSeatResponse(
-        Long concertId,
-        LocalDateTime concertAt,
-        List<ConcertSeatInfo> seats
-) {
-}
