--- src/main/java/hhplus/concert/api/dto/response/PaymentResponse.java
@@ -1,10 +1,0 @@
-package hhplus.concert.api.dto.response;
-
-import hhplus.concert.common.type.PaymentStatus;
-
-public record PaymentResponse(
-        Long paymentId,
-        Long amount,
-        PaymentStatus paymentStatus
-) {
-}
