--- src/main/java/hhplus/concert/api/dto/response/QueueTokenResponse.java
@@ -1,10 +1,0 @@
-package hhplus.concert.api.dto.response;
-
-import java.time.LocalDateTime;
-
-public record QueueTokenResponse(
-        Long tokenId,
-        LocalDateTime createdAt,
-        LocalDateTime expiredAt
-) {
-}
