--- src/main/java/hhplus/concert/api/dto/response/ReservationResponse.java
@@ -1,16 +1,0 @@
-package hhplus.concert.api.dto.response;
-
-import hhplus.concert.common.type.ReservationStatus;
-
-import java.time.LocalDateTime;
-
-public record ReservationResponse(
-        Long reservationId,
-        Long concertId,
-        String concertName,
-        LocalDateTime concertAt,
-        int SeatNumber,
-        int SeatPrice,
-        ReservationStatus reservationStatus
-) {
-}
