+++ src/main/java/hhplus/concert/api/presentation/dto/ConcertScheduleInfo.java
@@ -1,0 +1,10 @@
+package hhplus.concert.api.presentation.dto;
+
+import java.time.LocalDateTime;
+
+public record ConcertScheduleInfo(
+        Long scheduleId,
+        LocalDateTime concertAt,
+        LocalDateTime reservationAt
+) {
+}
