+++ src/main/java/hhplus/concert/api/presentation/dto/ConcertSeatInfo.java
@@ -1,0 +1,11 @@
+package hhplus.concert.api.presentation.dto;
+
+import hhplus.concert.support.type.SeatStatus;
+
+public record ConcertSeatInfo(
+        Long seatId,
+        int seatNumber,
+        SeatStatus seatStatus,
+        int SeatPrice
+) {
+}
