+++ src/main/java/hhplus/concert/api/presentation/dto/request/BalanceChargeRequest.java
@@ -1,0 +1,6 @@
+package hhplus.concert.api.presentation.dto.request;
+
+public record BalanceChargeRequest(
+        Long amount
+) {
+}
