+++ src/main/java/hhplus/concert/api/presentation/dto/request/PaymentRequest.java
@@ -1,0 +1,6 @@
+package hhplus.concert.api.presentation.dto.request;
+
+public record PaymentRequest(
+        Long reservationId
+) {
+}
