+++ src/main/java/hhplus/concert/api/presentation/dto/request/ReservationRequest.java
@@ -1,0 +1,11 @@
+package hhplus.concert.api.presentation.dto.request;
+
+import hhplus.concert.support.type.ReservationStatus;
+
+public record ReservationRequest(
+        Long userId,
+        Long concertId,
+        Long scheduleId,
+        ReservationStatus status
+) {
+}
