+++ src/main/java/hhplus/concert/api/presentation/dto/response/BalanceResponse.java
@@ -1,0 +1,7 @@
+package hhplus.concert.api.presentation.dto.response;
+
+public record BalanceResponse(
+        Long userId,
+        Long currentAmount
+) {
+}
