+++ src/main/java/hhplus/concert/api/presentation/dto/response/ConcertScheduleResponse.java
@@ -1,0 +1,11 @@
+package hhplus.concert.api.presentation.dto.response;
+
+import hhplus.concert.api.presentation.dto.ConcertScheduleInfo;
+
+import java.util.List;
+
+public record ConcertScheduleResponse(
+        Long concertId,
+        List<ConcertScheduleInfo> events
+) {
+}
