+++ src/main/java/hhplus/concert/api/presentation/dto/response/PaymentResponse.java
@@ -1,0 +1,10 @@
+package hhplus.concert.api.presentation.dto.response;
+
+import hhplus.concert.support.type.PaymentStatus;
+
+public record PaymentResponse(
+        Long paymentId,
+        Long amount,
+        PaymentStatus paymentStatus
+) {
+}
