+++ src/main/java/hhplus/concert/api/presentation/dto/response/QueueTokenResponse.java
@@ -1,0 +1,9 @@
+package hhplus.concert.api.presentation.dto.response;
+
+import java.time.LocalDateTime;
+
+public record QueueTokenResponse(
+        String tokenId,
+        LocalDateTime createdAt
+) {
+}
