--- src/main/java/hhplus/concert/common/exception/NotFoundException.java
@@ -1,7 +1,0 @@
-package hhplus.concert.common.exception;
-
-public class NotFoundException extends RuntimeException{
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
