--- src/main/java/hhplus/concert/common/type/PaymentStatus.java
@@ -1,6 +1,0 @@
-package hhplus.concert.common.type;
-
-public enum PaymentStatus {
-    COMPLETED,
-    FAILED
-}
