--- src/main/java/hhplus/concert/common/type/QueueStatus.java
@@ -1,7 +1,0 @@
-package hhplus.concert.common.type;
-
-public enum QueueStatus {
-    WAIT,
-    ACTIVE,
-    EXPIRED
-}
