--- src/main/java/hhplus/concert/common/type/ReservationStatus.java
@@ -1,10 +1,0 @@
-package hhplus.concert.common.type;
-
-public enum ReservationStatus {
-    PENDING,
-    CONFIRMED,
-    CANCELLED,
-    COMPLETED,
-    EXPIRED,
-    FAILED
-}
