--- src/main/java/hhplus/concert/common/type/SeatStatus.java
@@ -1,6 +1,0 @@
-package hhplus.concert.common.type;
-
-public enum SeatStatus {
-    AVAILABLE,
-    UNAVAILABLE,
-}
