--- src/main/java/hhplus/concert/domain/concert/Seat.java
@@ -1,29 +1,0 @@
-package hhplus.concert.domain.concert;
-
-import hhplus.concert.common.type.SeatStatus;
-import hhplus.concert.domain.concert.models.ConcertSchedule;
-import jakarta.persistence.*;
-
-@Entity
-public class Seat {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "seat_id")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "concert_schedule_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT))
-    private ConcertSchedule concertSchedule;
-
-    @Column(name = "seat_number", nullable = false)
-    private int seatNumber;
-
-    @Column(name = "seat_price", nullable = false)
-    private int seat_price;
-
-    @Column(name = "status", nullable = false)
-    @Enumerated(EnumType.STRING)
-    private SeatStatus status;
-
-}
