+++ src/main/java/hhplus/concert/domain/user/repositories/UserRepository.java
@@ -1,0 +1,8 @@
+package hhplus.concert.domain.user.repositories;
+
+
+import hhplus.concert.domain.user.models.User;
+
+public interface UserRepository {
+    User findById(Long userId);
+}
