+++ src/main/java/hhplus/concert/infra/jpa/JpaUserRepository.java
@@ -1,0 +1,8 @@
+package hhplus.concert.infra.jpa;
+
+import hhplus.concert.domain.user.models.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaUserRepository extends JpaRepository<User, Long> {
+}
+
