+++ src/main/java/hhplus/concert/support/exception/NotFoundException.java
@@ -1,0 +1,7 @@
+package hhplus.concert.support.exception;
+
+public class NotFoundException extends RuntimeException{
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
