+++ src/main/java/hhplus/concert/support/type/PaymentStatus.java
@@ -1,0 +1,6 @@
+package hhplus.concert.support.type;
+
+public enum PaymentStatus {
+    COMPLETED,
+    FAILED
+}
