+++ src/main/java/hhplus/concert/support/type/QueueStatus.java
@@ -1,0 +1,7 @@
+package hhplus.concert.support.type;
+
+public enum QueueStatus {
+    WAIT,
+    ACTIVE,
+    EXPIRED
+}
