+++ src/main/java/hhplus/concert/support/type/ReservationStatus.java
@@ -1,0 +1,10 @@
+package hhplus.concert.support.type;
+
+public enum ReservationStatus {
+    PENDING,
+    CONFIRMED,
+    CANCELLED,
+    COMPLETED,
+    EXPIRED,
+    FAILED
+}
