+++ src/main/java/hhplus/concert/support/type/SeatStatus.java
@@ -1,0 +1,6 @@
+package hhplus.concert.support.type;
+
+public enum SeatStatus {
+    AVAILABLE,
+    UNAVAILABLE,
+}
